refactor code :-)
