--- src/game_hooks.hpp
@@ -1,12 +1,0 @@
-#pragma once
-#ifndef _GAME_HOOKS_HPP
-#define _GAME_HOOKS_HPP
-
-#include <string>
-
-#include <MinHook.h>
-#include <cocos2d.h>
-
-void do_hooks();
-
-#endif
